fix: fixed the toggle in overwrite existing screen and fixed add to existing contact hooks
